+++ Game-cpp/CountdownTimer.h
@@ -17,7 +17,7 @@ public:
     void Start(); // Starts the timer
     void Restart();
     void Stop(); // Stops the timer
-    float UpdateTimer(int hiscore, float secondsPerLevel, bool drawTimer);
+    float UpdateTimer(uint16_t hiscore, float secondsPerLevel, bool drawTimer);
     float UpdateTimer(float secondsPerLevel);
     int timeInSeconds;
 
@@ -25,5 +25,5 @@ private:
     bool running;
     std::chrono::high_resolution_clock::time_point start;
 
-    void DrawTimer(int hiscore, double countdown, int secondsPerLevel);
+    void DrawTimer(uint16_t hiscore, float countdown, uint8_t secondsPerLevel);
 };
